WASM Compatibility (no bindings)
When building for WASM targets we don't want to generate the bindings
As compiling the project e.g. via wasm-pack, complains "stddef.h" isn't found.

We solve this by a feature flag "c_bindings", which will be enabled on non-wasm
targets by default, thus when included in a project being built for wasm it
will compile and work as expected when we use the library directly.

I also flagged the timer code in graph.rs as that's not compatible with wasm.